Fix the issues with migrating the GRFusion code. Now it is runnable with graph view construction DDL, and DQL like: select * from gv.vertexes
Some of DQLs are not originally supported by GRFusion1.0, like:
select v.* from gv.vertexes v
AND
select v.id from gv.vertexes v

These kinds of DQLs are supported in later commits which support
labeling of the vertices and edges.
This commit is only used for testing performance of the original
GRFusion1.0.